drivers: lmh-dcvs: Update cooling device registration and register space
Update the cooling device registration path to be done in a workqueue
instead of doing it in the CPU online callback. The registration path
could acquire mutex and could sleep, which makes it not suitable for
doing it in the online path.

Update the driver to queue a work in the CPU online path, which will
register for the CPU cooling device. Update the driver to specify
the OSM and LLM base address in the device tree and to remove the rcu
lock for OPP API calls.

Change-Id: I9ad0b90b7e340db05b2bdb2b584ea5b87bb8bfa8
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: DennySPB <dennyspb@gmail.com>
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>